feat: 멀티포인트 api호출을 위해 패키지형태로 구조변경
