+++ moodi/utils/config.py
@@ -1,0 +1,4 @@
+from dotenv import load_dotenv
+
+def load_config(env_path: str):
+    load_dotenv(env_path)
