+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "cat-holic-ai"
+name = "moodi"
 version = "0.1.0"
 description = ""
 authors = [
@@ -22,7 +22,8 @@ requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-package-mode = false
+package-mode = true
+packages = [{ include = "moodi" }]
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.5"
