+++ test/test_api.py
@@ -1,16 +1,14 @@
 from fastapi.testclient import TestClient
-from CAT_HOLIC_AI.main import app  # main.py의 FastAPI 인스턴스를 import
-from CAT_HOLIC_AI.utils import load_config 
+from moodi.main import app
+from moodi.utils.config import load_config  # 변경한 폴더명(utils)에 맞게 import 수정
 
-
-# 테스트 실행 전 환경변수 로딩
 load_config(env_path="config/dev.env")
 
 client = TestClient(app)
 
 def test_generate_story():
     response = client.post(
-        "/generate-story",
+        "/gemini/generate-story", 
         json={
             "age": 10,
             "social_situation": "새로운 학년이 시작되어 자기소개를 해야 하는 상황",
@@ -20,7 +18,7 @@ def test_generate_story():
 
     assert response.status_code == 200
     result = response.json()
-    print(result)  # 결과 확인을 위해 출력
+    print(result)
 
     assert "story" in result
     assert "question" in result
